+++ ui/package.json
@@ -78,6 +78,7 @@
     "react-datepicker": "7.5.0",
     "react-dom": "18.3.1",
     "react-i18next": "15.1.0",
+    "react-map-gl": "7.1.7",
     "react-resizable-panels": "2.1.6",
     "react-yjs": "2.0.1",
     "use-resize-observer": "9.1.0",
