+++ ui/src/components/visualizations/index.ts
@@ -1,0 +1,2 @@
+export * from "./ThreeDMap";
+export * from "./TwoDMap";
