+++ ui/src/features/Editor/components/BottomPanel/components/Map/index.tsx
@@ -1,8 +1,4 @@
-import { CoreVisualizer, SceneProperty, Layer } from "@reearth/core";
-import { useEffect, useState } from "react";
-
-import { Button, TwoDMap } from "@flow/components";
-import fires from "@flow/mock_data/fires.json";
+import { Button, ThreeDMap, TwoDMap } from "@flow/components";
 
 export type MapMode = "2d" | "3d";
 
@@ -14,55 +10,6 @@ type Props = {
 };
 
 const Map: React.FC<Props> = ({ mapMode, setMapMode }) => {
-  const [isReady, setIsReady] = useState(false);
-
-  const sceneProperty: SceneProperty = {
-    default: {
-      sceneMode: mapMode,
-    },
-    camera: {
-      camera: {
-        lng: 127.05177672074426,
-        lat: -6.260283141094241,
-        height: 7594277.78896907,
-        heading: 1.129814464206902e-9,
-        pitch: -1.5707963267948966,
-        roll: 0,
-        fov: 1,
-        aspectRatio: 1,
-      },
-    },
-    tiles: [
-      {
-        id: "default",
-        tile_type: "default",
-      },
-    ],
-  };
-
-  const layers: Layer[] = [
-    {
-      id: "marker",
-      type: "simple",
-      data: {
-        type: "geojson",
-        value: fires,
-      },
-      marker: {
-        imageColor: {
-          expression: {
-            conditions: [["true", "color('#FF0000')"]],
-          },
-        },
-      },
-    },
-  ];
-
-  useEffect(() => {
-    if (isReady) return;
-    setIsReady(true);
-  }, [isReady]);
-
   return (
     <div className="flex w-1/2">
       <div className="relative w-full">
@@ -78,17 +25,7 @@ const Map: React.FC<Props> = ({ mapMode, setMapMode }) => {
             </Button>
           ))}
         </div>
-        {mapMode === "2d" ? (
-          <TwoDMap />
-        ) : (
-          <CoreVisualizer
-            engine="cesium"
-            isBuilt
-            ready={isReady}
-            sceneProperty={sceneProperty}
-            layers={layers}
-          />
-        )}
+        {mapMode === "2d" ? <TwoDMap /> : <ThreeDMap />}
       </div>
     </div>
   );
